parsing written need check

+++ lib/libft/inc/libft.h
@@ -48,7 +48,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t n);
 void	*ft_memchr(const void *s, int c, size_t n);
 void	*ft_calloc(size_t nmemb, size_t size);
-void	*ft_realloc(void *ptr, int new_size, int old_size);
+void	*ft_realloc(void *ptr, size_t new_size, size_t old_size);
 char	*ft_strchr(const char *s, int c);
 char	*ft_strnstr(const char *big, const char *small, size_t len);
 char	*ft_strdup(const char *s);
